requirement.txt
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
Django==5.1.2
django-cors-headers==4.6.0
django-jazzmin==3.0.1
django-minio-storage==0.5.7
django-ninja==1.3.0
django-quill-editor==0.1.42
django-storages==1.14.4
dnspython==2.7.0
email_validator==2.2.0
idna==3.10
minio==7.2.10
pillow==11.0.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.8.2
